refactor(blocks): entities/block を entities/blocks-field にリネームし命名規約に整合させる (#501)
フォルダ内の全シンボルは BlocksField*（OpenAPI タグ BlocksFields / /api/v1/blocks-fields / query-key ['blocks-fields']）なのにフォルダ名だけ単数 block で frontend-standards のリソース命名（{type}-field/）に違反していた。git mv で blocks-field へ揃え、importer 2箇所を更新する。シンボル名・挙動は不変。
